fix: fix savannah errors occured after moving logic to service, move Answer model outside of ISprint model
