+++ src/app/core/models/IAnswer.ts
@@ -1,0 +1,4 @@
+export enum Answer {
+  CORRECT = 'CORRECT',
+  WRONG = 'WRONG',
+}
