+++ src/app/mini-games/shared/components/end-game/end-game.component.spec.ts
@@ -3,9 +3,8 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-
 import { EndGameComponent } from './end-game.component';
-import { Answer } from 'src/app/core/models/ISprintGame';
+import { Answer } from 'src/app/core/models/IAnswer';
 
 describe('EndGameComponent', () => {
   let component: EndGameComponent;
